refactor: Consolidate CommandRegistry and CommandGroup into sectioned_group.py
